refactor: split Parlant tools into a package with shared helpers
Convert the monolithic parlant/tools.py into a tools/ package (registry,
helpers, chat, contacts, memory, factory) and collapse the per-tool session
guard, error handling, model normalization, and JSON serialization into shared
helpers. Public API and tool behavior are unchanged.
